refactor(journey-types): rename to factory and remove switch
